requirements-dev.txt
thunder-python
numpy
scipy
future
pytest
pytest-rerunfailures
pytest-timeout
pytest-mock
mock